pyproject.toml
[tool.poetry]
name = "ncovenience"
version = "1.1.0"
description = ""
authors = ["Kenneth V. Domingo <hello@kvdomingo.xyz>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.11.1"
geojson = "^2.5.0"
gunicorn = {extras = ["gthread"], version = "^20.1.0"}
numpy = "^1.23.2"
pandas = "^1.4.3"
Pillow = "^9.2.0"
python-dotenv = "^0.20.0"
requests = "^2.28.1"
whitenoise = "^6.2.0"
Flask = "^2.2.2"
Flask-Caching = "^2.0.1"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py310']
line-length = 120